timestamps & multiple graphics drawing fix & graph color theme update

+++ app/src/main/java/ru/kremlev/ncsrecognitonmanager/manager/data/TimeStamp.kt
@@ -1,5 +1,5 @@
 package ru.kremlev.ncsrecognitonmanager.manager.data
 
 data class TimeStamp(
-    var ms: Int
+    var ms: Long
 )
